backend/pyproject.toml
[tool.poetry]
name = "efiko-backend"
version = "0.1.0"
description = "Backend for Efiko AI Chatbot"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
supabase = "^1.0.3"
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
google-generativeai = "^0.3.1"
numpy = "^1.24.3"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
